Cast variable to int to give sprintf the correct type

+++ src/eez/mp.cpp
@@ -339,7 +339,7 @@ bool scpi(const char *commandOrQueryText, const char **resultText, size_t *resul
 
     if (g_lastError != 0) {
         static char g_scpiError[48];
-        snprintf(g_scpiError, 48, "SCPI error %d, \"%s\"", g_lastError, SCPI_ErrorTranslate(g_lastError));
+        snprintf(g_scpiError, 48, "SCPI error %d, \"%s\"", (int)g_lastError, SCPI_ErrorTranslate(g_lastError));
         mp_raise_ValueError(g_scpiError);
     }
